event 模块 IOCP 事件循环对齐 Unix 模型 注册与连接投递移入 watcher 线程，移除 overlap_tcp_ctx 的 conn_addr 字段改由命令透传，conn/fd/lsn inloop 去除冗余 fd 参数
修复协议与服务层的边界及安全问题并补测
RESP/UDP 长度头位宽、DataCenter key 长度校验、随机数失败处理、序列化锁唤醒状态恢复、Lua 绑定参数校验

构建 LTO 按编译器区分并行度
GCC 用 -flto=核数消除 lto-wrapper 串行编译警告

统一初始化 { 0 } 写法与结构体字段对齐

+++ lib/protocol/redis.c
@@ -55,7 +55,7 @@ static inline void _redis_create_sds(binary_ctx *fbuf, binary_ctx *sdsbuf, size_
     if (0 == fbuf->offset) {
         return;
     }
-    binary_set_va(sdsbuf, "$%u"FLAG_CRLF, (uint32_t)fbuf->offset);
+    binary_set_va(sdsbuf, "$%zu"FLAG_CRLF, fbuf->offset);
     binary_set_string(sdsbuf, fbuf->data, fbuf->offset);
     binary_set_string(sdsbuf, FLAG_CRLF, CRLF_SIZE);
     binary_offset(fbuf, 0);
